snip most cmdutil deps
