--- pkg/cmd/util/doc.go
@@ -1,2 +1,0 @@
-// Package util provides utility functions for the cmd packages.
-package util
