+++ pkg/cmd/util/net.go
@@ -1,124 +1,10 @@
 package util
 
 import (
-	"crypto/tls"
-	"net"
-	"net/http"
 	"strings"
-	"time"
 
 	"k8s.io/apimachinery/pkg/util/sets"
-
-	"k8s.io/klog"
 )
-
-// TryListen tries to open a connection on the given port and returns true if it succeeded.
-func TryListen(network, hostPort string) (bool, error) {
-	l, err := net.Listen(network, hostPort)
-	if err != nil {
-		klog.V(5).Infof("Failure while checking listen on %s: %v", hostPort, err)
-		return false, err
-	}
-	defer l.Close()
-	return true, nil
-}
-
-// tcpKeepAliveListener sets TCP keep-alive timeouts on accepted
-// connections. It's used by ListenAndServe and ListenAndServeTLS so
-// dead TCP connections (e.g. closing laptop mid-download) eventually
-// go away.
-type tcpKeepAliveListener struct {
-	*net.TCPListener
-}
-
-func (ln tcpKeepAliveListener) Accept() (c net.Conn, err error) {
-	tc, err := ln.AcceptTCP()
-	if err != nil {
-		return
-	}
-	tc.SetKeepAlive(true)
-	tc.SetKeepAlivePeriod(3 * time.Minute)
-	return tc, nil
-}
-
-// ListenAndServeTLS starts a server that listens on the provided TCP mode (as supported
-// by net.Listen).
-func ListenAndServeTLS(srv *http.Server, network string, certFile, keyFile string) error {
-	addr := srv.Addr
-	if addr == "" {
-		addr = ":https"
-	}
-	config := &tls.Config{}
-	if srv.TLSConfig != nil {
-		config = srv.TLSConfig
-	}
-	if config.NextProtos == nil {
-		config.NextProtos = []string{"http/1.1"}
-	}
-
-	var err error
-	config.Certificates = make([]tls.Certificate, 1)
-	config.Certificates[0], err = tls.LoadX509KeyPair(certFile, keyFile)
-	if err != nil {
-		return err
-	}
-
-	ln, err := net.Listen(network, addr)
-	if err != nil {
-		return err
-	}
-
-	tlsListener := tls.NewListener(tcpKeepAliveListener{ln.(*net.TCPListener)}, config)
-	return srv.Serve(tlsListener)
-}
-
-// WaitForSuccessfulDial attempts to connect to the given address, closing and returning nil on the first successful connection.
-func WaitForSuccessfulDial(https bool, network, address string, timeout, interval time.Duration, retries int) error {
-	var (
-		conn net.Conn
-		err  error
-	)
-	for i := 0; i <= retries; i++ {
-		dialer := net.Dialer{Timeout: timeout}
-		if https {
-			conn, err = tls.DialWithDialer(&dialer, network, address, &tls.Config{InsecureSkipVerify: true})
-		} else {
-			conn, err = dialer.Dial(network, address)
-		}
-		if err != nil {
-			klog.V(5).Infof("Got error %#v, trying again: %#v\n", err, address)
-			time.Sleep(interval)
-			continue
-		}
-		conn.Close()
-		return nil
-	}
-	return err
-}
-
-// GetCertificateFunc returns a function that can be used in tls.Config#GetCertificate
-// Returns nil if len(certs) == 0
-func GetCertificateFunc(certs map[string]*tls.Certificate) func(*tls.ClientHelloInfo) (*tls.Certificate, error) {
-	if len(certs) == 0 {
-		return nil
-	}
-	// Replica of tls.Config#getCertificate logic
-	return func(clientHello *tls.ClientHelloInfo) (*tls.Certificate, error) {
-		if clientHello == nil {
-			return nil, nil
-		}
-
-		name := clientHello.ServerName
-		name = strings.ToLower(name)
-		name = strings.TrimRight(name, ".")
-		for _, candidate := range HostnameMatchSpecCandidates(name) {
-			if cert, ok := certs[candidate]; ok {
-				return cert, nil
-			}
-		}
-		return nil, nil
-	}
-}
 
 // HostnameMatchSpecCandidates returns a list of match specs that would match the provided hostname
 // Returns nil if len(hostname) == 0
